backend/src/backend/requirements.txt
numpy==1.26.4
django==3.1.1
celery[redis]==5.0.5 
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
gunicorn==20.0.4
idna==2.10
msgpack==1.0.2
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3 
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
wand==0.6.6
requests==2.25.1
django-cors-headers==3.7.0
librosa==0.10.1
imageio[pyav,ffmpeg]==2.33.1
grpcio==1.60.0
grpcio-tools==1.60.0
pympi-ling==1.70.2
pandas==2.2.2
ffmpeg-python
msgpack_numpy